+++ drivers/pinctrl/Kconfig.mcux
@@ -2,10 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 DT_COMPAT_MCUX_RT_PINCTRL := nxp,mcux-rt-pinctrl
+DT_COMPAT_MCUX_RT11XX_PINCTRL := nxp,mcux-rt11xx-pinctrl
 
 config PINCTRL_MCUX_RT
 	bool "Pin controller driver for MCUX RT1xxx MCUs"
 	depends on SOC_SERIES_IMX_RT
-	default $(dt_compat_enabled,$(DT_COMPAT_MCUX_RT_PINCTRL))
+	default $(dt_compat_enabled,$(DT_COMPAT_MCUX_RT_PINCTRL)) ||	\
+		$(dt_compat_enabled,$(DT_COMPAT_MCUX_RT11XX_PINCTRL))
 	help
 	  Enable pin controller driver for NXP RT series MCUs
